+++ src/actions/updateIndex.js
@@ -3,3 +3,9 @@ export const updateIndex = index => ({
   type: UPDATE_INDEX,
   payload: index
 });
+
+export const UPDATE_BODY = 'UPDATE_BODY';
+export const updateDocumentBody = body => ({
+  type: UPDATE_BODY,
+  payload: body
+});
