+++ src/containers/MarkdownList.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import List from '../components/markdown/List';
-import { getMarkdownList } from '../selectors/markdownListSelector';
+import { getMarkdownList } from '../selectors/markdownSelector';
 
 const mapStateToProps = state => ({
   markdownFiles: getMarkdownList(state)
