--- src/selectors/markdownDocumentSelector.js
@@ -1,3 +1,0 @@
-export const getDocument = state => {
-  return state.markdownReducer.documents[state.markdownReducer.index];
-};
