--- src/selectors/markdownListSelector.js
@@ -1,3 +1,0 @@
-export const getMarkdownList = state => {
-  return state.markdownReducer.documents;
-};
